backend/requirements.txt
Flask
python-dotenv
flask-restful
flask_sqlalchemy
flask-jwt-extended